Muuuuuuuuuuita coisa pra caber aqui
- ExtendedWebViewRenderer: texto justificado (aplicado nas telas "Sobre" e "Detalhes");
- Limite de caracteres em campos de peso e altura;
- Refatoração completa de notificações (correção de bugs + melhoria no fluxo de funcionamento do código);
- Bug de informações aparecendo no lugar errado na timeline principal corrigido;
- Override de back nas paginas principais p/ voltar para a pagina inicial do app;
- Clique na foto do Drawer volta para a pagina inicial;
- Ao clicar em qualquer botão de formulário, o aplicativo entra em loading instantaneamente e evita múltiplos cliques;
- Melhoria na validação + remoção de emojis em campos de texto; e
- Otimização de código geral.

+++ ProMama/ProMama/Models/Notificacao.cs
@@ -5,7 +5,7 @@
         public int id { get; set; }
         public string titulo { get; set; }
         public string texto { get; set; }
-        public int semana { get; set; }
+        public double semana { get; set; }
 
         public Notificacao() { }
     }
